OXDEV-9653 Improve API entrypoint codeception tests
* Move api entrypoint tests to separate directory
* Improved readability of JS calls
* Removed some tests that covers the same stuff as we have already covered by integrations
